perf(elements): ⚡️ Improve generating performance and streamlining
This commit is a bit late and comes after a lot of tests. It should add: an updatequeue, with accompanying changes in the sync element.update function. It is checked on multiple locations, and emptied via the async command to prevent race conditions; Selector now has a skip_update attribute, which allows the select element to not update when selecting a new element (i.e. if that is done in the on_select function, this prevent  multiple generator runs)