feat(elements): 🚸 make GridLayout columns and rows default to None
After using, it felt unintuitive to have to specify one as None when you would want to have a set number of rows/columns. This way, defining only one of the two will have the other value at None. If both are undefined, the previous default (4x4) is silently applied.