fix(styles): 🐛 Fix colorproperty calling the fset of a previous iteration
The fset function that was passed via the getter would not be the one from the class, but the one from the object the getter was set to. This caused the fset function that was called to have the 'self' be an outdated property, who's __set_name__ function would not be called.